Rename custom Trello typings and improve theDm
